Added initial backend server, Updated models to reflect new requirements Updated launch scripts to reflect new server
