+++ group5/package.json
@@ -1,27 +1,32 @@
 {
-  "name": "group5",
-  "version": "0.0.0",
-  "scripts": {
-    "dev": "vite",
-    "build": "vite build",
-    "serve": "vite preview",
-    "typeorm": "node --require ts-node/register ./node_modules/typeorm/cli.js"
-  },
-  "dependencies": {
-    "@vitejs/plugin-vue-jsx": "^1.1.0",
-    "mysql2": "^2.2.5",
-    "reflect-metadata": "^0.1.13",
-    "typeorm": "^0.2.31",
-    "vue": "^3.0.5",
-    "vue-router": "^4.0.4"
-  },
-  "devDependencies": {
-    "@types/node": "^14.14.31",
-    "@vitejs/plugin-vue": "^1.1.4",
-    "@vue/compiler-sfc": "^3.0.5",
-    "esbuild": "^0.8.51",
-    "ts-node": "^9.1.1",
-    "typescript": "^4.1.3",
-    "vite": "^2.0.1"
-  }
+    "name": "group5",
+    "version": "0.0.0",
+    "scripts": {
+        "dev": "(vite) & nodemon --watch ./src/server/**/*.ts --exec ts-node ./src/server/app.ts",
+        "front": "vite",
+        "back": "nodemon --watch ./src/server/**/*.ts --exec ts-node ./src/server/app.ts",
+        "build": "vite build",
+        "serve": "vite preview",
+        "typeorm": "node --require ts-node/register ./node_modules/typeorm/cli.js"
+    },
+    "dependencies": {
+        "@types/express": "^4.17.11",
+        "@vitejs/plugin-vue-jsx": "^1.1.0",
+        "express": "^4.17.1",
+        "mysql2": "^2.2.5",
+        "reflect-metadata": "^0.1.13",
+        "typeorm": "^0.2.31",
+        "vue": "^3.0.5",
+        "vue-router": "^4.0.4"
+    },
+    "devDependencies": {
+        "@types/node": "^14.14.31",
+        "@vitejs/plugin-vue": "^1.1.4",
+        "@vue/compiler-sfc": "^3.0.5",
+        "esbuild": "^0.8.51",
+        "nodemon": "^2.0.7",
+        "ts-node": "^9.1.1",
+        "typescript": "^4.1.3",
+        "vite": "^2.0.1"
+    }
 }
