+++ group5/src/server/entity/Reservation.ts
@@ -1,0 +1,64 @@
+// TypeORM imports
+import {
+    Entity,
+    Column,
+    PrimaryGeneratedColumn,
+    CreateDateColumn,
+    UpdateDateColumn,
+    DeleteDateColumn,
+    ManyToOne,
+    JoinColumn,
+    ManyToMany
+} from "typeorm";
+
+// Model imports
+import { User } from "./User";
+import { Room } from "./Room";
+
+// Reservation Model
+// Contains all of the reservations in the system
+@Entity()
+export class Reservation {
+
+    // TypeORM generated UUID for uniqueness
+    @PrimaryGeneratedColumn("uuid")
+    uuid!: string;
+
+    // Start date of the reservation
+    @Column()
+    startDate!: Date;
+
+    // End date of the reservation
+    @Column()
+    endDate!: Date;
+
+    // Room assigned to the reservation
+    @ManyToOne(() => Room, assignedRoom => assignedRoom.reservations)
+    assignedRoom!: Room;
+
+    // User who the reservation is for
+    @ManyToOne(() => User, createdBy => createdBy.reservations)
+    createdBy!: User;
+
+    // If the admin created the reservation
+    @Column({ default: false })
+    adminCreate!: boolean;
+
+    // Save any special requests
+    @Column({ default: "None" })
+    specialRequests!: string;
+
+    // Auto generated by TypeORM for the initial creation date
+    @CreateDateColumn()
+    created!: Date;
+
+    // Auto generated by TypeORM for the last updated date
+    @UpdateDateColumn()
+    updated!: Date;
+
+    // Auto generated by TypeORM for the deletion (canx) date
+    @DeleteDateColumn()
+    deleted!: Date;
+
+
+}
