+++ group5/src/server/entity/RoomType.ts
@@ -1,0 +1,41 @@
+// TypeORM imports
+import {
+    Entity,
+    Column,
+    PrimaryGeneratedColumn,
+    CreateDateColumn,
+    UpdateDateColumn,
+    DeleteDateColumn,
+    ManyToOne
+} from "typeorm";
+
+// The RoomType model defines a specific room type
+// There can be multiple rooms with the same type
+@Entity()
+export class RoomType {
+
+    // UUID auto generated by TypeORM for uniqueness
+    @PrimaryGeneratedColumn("uuid")
+    uuid!: string;
+
+    // Name of the room type, must be unique
+    @Column({ unique: true })
+    typeName!: string;
+
+    // The cost of the room
+    @Column()
+    typeCost!: number;
+
+    // Auto generated by TypeORM when record is created
+    @CreateDateColumn()
+    created!: Date;
+
+    // Auto generated by TypeORM when the record is modified
+    @UpdateDateColumn()
+    updated!: Date;
+
+    // Auto generated by TypeORM when the record is deleted
+    @DeleteDateColumn()
+    deleted!: Date;
+
+}
